+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `name()` method to cache files for retrieving the file name.
+
 ## [0.0.1] - 2025-08-29
 
 ### Added
@@ -21,5 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
+[Unreleased]: https://github.com/ventaquil/fcache/compare/v0.0.1...HEAD
 [0.0.1]: https://github.com/ventaquil/fcache/compare/v0.0.0...v0.0.1
 [0.0.0]: https://github.com/ventaquil/fcache/releases/tag/v0.0.0